requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-google-genai
google-generativeai
pypdf
sentence-transformers
matplotlib
numpy
pandas
faiss-cpu
python-dotenv
unstructured
libmagic
python-magic
python-magic-bin
seaborn
